bugfix
Need to check user edited phrases for errors such as extra spaces and tabs.
User edited text in table need to be extracted from the 'edited  formatted field' when the data source is not refreshed.  The phrases dialog box was not correctly obtaining the checkbox value when phrases were removed.  The phrases are now showing the "checkbox" information.